+++ CHANGELOG.md
@@ -6,11 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+* Modified `es.spatial()` to include a function called `crop_all()` that crops a list of images (@nkorinek, #333) 
 * Updated stack to ensure inputs are of same rows/cols/bands, CRS, and Affine transform (@joemcglinchy, #334) 
 * Add vignette for using the plot_bands() function (@nkorinek, #315)
 * Add example vignette for stacking and cropping raster bands with EarthPy (@nkorinek, #267)
 * Add example vignette for plotting band combinations with EarthPy (@jlpalomino, #306)
-* Modified `es.spatial()` to include a function called `crop_all()` that crops a list of images (@nkorinek, #333) 
 
 ## [0.6.9]
 * Add URL to io.py for Landsat vignette dataset (@jpalomino, #309)